+++ test/Language/JVM/ByteCodeSpec.hs
@@ -7,32 +7,48 @@ import Data.Either
 -- containers
 import qualified Data.IntMap as IntMap
 
-import Language.JVM.ByteCode
 import Language.JVM.ClassFileReader
 import Language.JVM.ConstantPool
 import Language.JVM.Constant
+import Language.JVM.ByteCode
 
 spec :: Spec
 spec = do
-  describe "new RefType" $ do
-    it "should evolve New" $ do
+  fdescribe "new RefType" $ do
+    it "should evolve 1, [J -> [[J" $ do
         let cp = ConstantPool (IntMap.singleton 0 (CClassRef "[J"))
         let x = runEvolve (EvolveConfig [] cp (const True))
-                (evolveBC undefined (ByteCodeInst 0 (New (Reference 0 1))))
-        x `shouldBe` Right (ByteCodeInst 0 (New "[J"))
+                (evolveBC undefined
+                 (ByteCodeInst 0 (NewArray (ArrayReference 0 1)))
+                )
+        x `shouldBe` Right (ByteCodeInst 0 (NewArray (NewArrayType 1 "[J")))
 
-    it "should evolve New" $ do
-        let cp = ConstantPool (IntMap.singleton 0 (CClassRef "java/lang/Object"))
+    it "should evolve 2, [[J -> [[J" $ do
+        let cp = ConstantPool (IntMap.singleton 0 (CClassRef "[[J"))
         let x = runEvolve (EvolveConfig [] cp (const True))
-                (evolveBC undefined (ByteCodeInst 0 (New (Reference 0 1))))
-        x `shouldBe` Right (ByteCodeInst 0 (New "[Ljava/lang/Object;"))
+                (evolveBC undefined
+                 (ByteCodeInst 0 (NewArray (ArrayReference 0 2)))
+                )
+        x `shouldBe` Right (ByteCodeInst 0 (NewArray (NewArrayType 2 "J")))
 
-    it "should roundtrip New" $ do
-        let x = byteCodeRoundtrip (ByteCodeInst 0 (New "[J"))
+    -- it "should evolve New" $ do
+    --     let cp = ConstantPool (IntMap.singleton 0 (CClassRef "J"))
+    --     let x = runEvolve (EvolveConfig [] cp (const True))
+    --             (evolveBC undefined (ByteCodeInst 0 (NewArray (ArrayReference 0 1))))
+    --     x `shouldBe` Right (ByteCodeInst 0 (NewArray (NewArrayType 1 "[J"))
+
+    -- it "should evolve New" $ do
+    --     let cp = ConstantPool (IntMap.singleton 0 (CClassRef "java/lang/Object"))
+    --     let x = runEvolve (EvolveConfig [] cp (const True))
+    --             (evolveBC undefined (ByteCodeInst 0 (New (Reference 0 1))))
+    --     x `shouldBe` Right (ByteCodeInst 0 (New "[Ljava/lang/Object;"))
+
+    it "should roundtrip NewArray" $ do
+        let x = byteCodeRoundtrip (ByteCodeInst 0 (NewArray (NewArrayType 1 "J")))
         x `shouldSatisfy` isRight
 
-    it "should roundtrip New" $ do
-        let x = byteCodeRoundtrip (ByteCodeInst 0 (New "[Ljava/lang/Object;"))
+    it "should roundtrip NewArray" $ do
+        let x = byteCodeRoundtrip (ByteCodeInst 0 (NewArray (NewArrayType 2 "J")))
         x `shouldSatisfy` isRight